build(repo): PR suggestions
